refactor (ListFragmentActivity): implementando RxJava
