+++ app/src/main/java/com/testemobile/githubjava/NetWork/PullRequestEndpoint.kt
@@ -1,6 +1,7 @@
-package com.testemobile.githubjava.Retrofit
+package com.testemobile.githubjava.NetWork
 
 import com.google.gson.JsonArray
+import io.reactivex.Maybe
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -11,6 +12,6 @@ interface PullRequestEndpoint {
     fun getPullRequest(
         @Path(value = "criador")criador :String,
         @Path(value = "repositorio") repositorio:String
-    ):Call<JsonArray>
+    ): Maybe<JsonArray>
 
 }
