+++ app/src/main/java/com/testemobile/githubjava/NetWork/RequestRepoEndpoint.kt
@@ -1,7 +1,7 @@
-package com.testemobile.githubjava.Retrofit
+package com.testemobile.githubjava.NetWork
 
 import com.testemobile.githubjava.Model.GitHubRepo
-import retrofit2.Call
+import io.reactivex.Maybe
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -15,6 +15,6 @@ interface RequestRepoEndpoint {
     }
 
     @GET(REPO_ENDPOINT)
-    fun getItems(@Query("page=1") numberPage: String?) : Call<GitHubRepo>
+    fun getItems(@Query("page=1") numberPage: String?) : Maybe<GitHubRepo>
 
 }
